+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelogs
 
+### v0.9.0
+
+- Implemented ReplayGain 2.0 support
+- Added master gain setting for extra output volume
+- Fixed a small PKGBUILD issue
+
 ### v0.8.1
 
 - Few project changes